+++ lib/src/web/verisoul_sdk_plugin_web.dart
@@ -14,7 +14,8 @@ class VerisoulSdkPlugin extends VerisoulApiHostApi {
   }
 
   @override
-  Future<void> configure(int enviromentVariable, String projectId) async {}
+  Future<void> configure(
+      int enviromentVariable, String projectId, bool reinitialize) async {}
 
   /// ✅ Check if the Verisoul SDK is loaded
   bool isVerisoulLoaded() {